+++ cmake/ThirdPartyBoost.cmake
@@ -10,7 +10,7 @@
 MESSAGE(STATUS "Searching for Boost:")
 SET(MIN_VER "1.56.0")
 SET(NEEDED_BOOST_LIBS thread iostreams date_time filesystem system
-    program_options regex timer chrono)
+    program_options regex)
 SET(Boost_DEBUG 0)
 SET(Boost_NO_BOOST_CMAKE ON)
 IF( BOOST_ROOT )