+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v5.0.0
 - Significant overhaul of CMake infrastructure (!770, !804)
 - Fix ThridpartyCCM options (!802)
 - Fix Windows CRLF tokens in GEO reader and improve comment handling (!805)
+- Use chrono in Timer (!807)
 
 **NekMesh**:
 - Add feature to read basic 2D geo files as CAD (!731)